% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sample_prob_polyA.R
\name{calculate_sample_prob_polyA}
\alias{calculate_sample_prob_polyA}
\title{given a transcript and fragment length, calculate probability of generating reads along the transcript}
\usage{
calculate_sample_prob_polyA(transcript, fraglen, polyAnum = 15,
  bias = "empirical", model = NULL)
}
\arguments{
\item{transcript}{transcript sequence, DNAString.}

\item{fraglen}{fragment length, integer.}

\item{polyAnum}{minimum number of 'A's in a polyA region, integer. A region should contain at least n continous 'A's to be considered as a polyA region.}

\item{bias}{polyA sampling bias, one of 'empirical', 'custom', 'naive' (default 'empirical'). If 'empirical', the built-in model is adopted.
This model is trained on data sets from the Drop-seq experiments described in the Cell paper (Macosko et al, 2015).
If 'custom', polyA sampling bias is captured from user input model. The path of user input model is specified in \code{model}.
If 'naive', sampling probability is calculated as (1/distance to polyA region).}

\item{model}{path to polyA sampling bias model if \code{bias} set to 'custom'. Model format should be a data frame with 2 columns.
Column 1 is distance to polyA region, while Column 2 is probability of getting a fragment at that position. Column 2 sums to 1.}
}
\value{
probability of generating reads along the transcript
}
\description{
given a transcript and fragment length, calculate probability of generating reads along the transcript
}
