% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reverse_complement.R
\name{find_reverse_complement}
\alias{find_reverse_complement}
\title{find reverse complement sequence given a transcript}
\usage{
find_reverse_complement(transcript)
}
\arguments{
\item{transcript}{transcript sequence, DNAString.}
}
\value{
reverse complement sequence of input transcript, DNAString.
}
\description{
find reverse complement sequence given a transcript
}
