% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barcodes.R
\name{generate_barcodes}
\alias{generate_barcodes}
\title{generate barcodes
generate cell barcode and molecolar barcode for each read}
\usage{
generate_barcodes(dirpath)
}
\arguments{
\item{dirpath}{file path specifying where sequencing reads should be written}
}
\value{
No return, but simulated reads are written to \code{filepath}.
}
\description{
generate barcodes
generate cell barcode and molecolar barcode for each read
}
