% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_countmat.R
\name{generate_countmat}
\alias{generate_countmat}
\title{generate a countmat matrix given gene number and cell number}
\usage{
generate_countmat(ngenes, ncells, loc, scale)
}
\arguments{
\item{ngenes}{gene number}

\item{ncells}{cell number}

\item{loc}{location parameter for the library size log-normal distribution, together with \code{libscale} controls library size. Determine different sequencing depth (number of reads) between cells.}

\item{scale}{scale parameter for the library size log-normal distribution, together with \code{libloc} controls library size. Determine different sequencing depth (number of reads) between ceels.}
}
\value{
a countmat matrix
}
\description{
generate a countmat matrix given gene number and cell number
}
