% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sample_prob_polyA.R
\name{calculate_sample_prob_polyA}
\alias{calculate_sample_prob_polyA}
\title{given a transcript and fragment length, calculate probability of generating reads along the transcript}
\usage{
calculate_sample_prob_polyA(transcript, fraglen)
}
\arguments{
\item{transcript}{transcript sequence, DNAString.}

\item{fraglen}{fragment length, integer.}
}
\value{
probability of generating reads along the transcript
}
\description{
given a transcript and fragment length, calculate probability of generating reads along the transcript
}
