% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_polyA_regions.R
\name{find_polyA_regions}
\alias{find_polyA_regions}
\title{find polyA regions given a transcript}
\usage{
find_polyA_regions(transcript)
}
\arguments{
\item{transcript}{transcript sequence, DNAString.}
}
\value{
polyA regions along the transcript, list.
}
\description{
find polyA regions given a transcript
}
