% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_countmat.R
\name{generate_countmat}
\alias{generate_countmat}
\title{generate a countmat matrix given gene number and cell number}
\usage{
generate_countmat(ngenes, ncells)
}
\arguments{
\item{ngenes}{gene number}

\item{ncells}{cell number}
}
\value{
a countmat matrix
}
\description{
generate a countmat matrix given gene number and cell number
}
