% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dropseq_experiment.R
\name{simulate_dropseq_experiment}
\alias{simulate_dropseq_experiment}
\title{Simulator: simulate_dropseq_experiment
Simulate Dropseq RNA-seq reads}
\usage{
simulate_dropseq_experiment(fasta = NULL, ngenes = NULL,
  ncells = NULL, outdir = ".")
}
\arguments{
\item{fasta}{path to FASTA file containing transcripts from which to simulate reads.}

\item{ngenes}{number of genes to simulate}

\item{ncells}{number of cells to simulate}

\item{outdir}{path to folder where simulated reads should be written. By default, reads written to the working directory.}
}
\value{
No return, but simulated reads are written to \code{outdir}.
}
\description{
create FASTA files containing RNA-seq reads simulated from provided transcripts, given gene number and cell number
}
