% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dropseq_experiment_countmat.R
\name{simulate_dropseq_experiment_countmat}
\alias{simulate_dropseq_experiment_countmat}
\title{Simulator: simulate_dropseq_experiment_countmat
Simulate Dropseq RNA-seq reads}
\usage{
simulate_dropseq_experiment_countmat(fasta = NULL, readmat,
  outdir = ".")
}
\arguments{
\item{fasta}{path to FASTA file containing transcripts from which to simulate reads.}

\item{readmat}{a gene-cell matrix, each entry represents number of reads to simulate}

\item{outdir}{path to folder where simulated reads should be written. By default, reads written to the working directory.}
}
\value{
No return, but simulated reads are written to \code{outdir}.
}
\description{
create FASTA files containing RNA-seq reads simulated from provided transcripts, given a countmat matrix
}
